CLN: remove _pickle_array, _unpkckle_array from common.py
